web_app/web_app/requirements.txt
blinker==1.9.0
cairocffi==1.7.1
cairosvg==2.8.2
click==8.1.8
colorama==0.4.6
comtypes==1.4.11
cssselect2==0.8.0
decorator==5.2.1
defusedxml==0.7.1
Flask==3.1.1
flask-cors==6.0.1
gTTS==2.5.4
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
moviepy==2.2.1
numpy==2.2.6
pdf2image==1.17.0
pillow==11.3.0
proglog==0.1.12
pyttsx3==2.99
pyzbar==0.1.9
python-dotenv==1.1.1
qrcode==8.2
tinycss2==1.4.0
tqdm==4.67.1
webencodings==0.5.1
Werkzeug==3.1.3
